no funca nada

+++ client/src/components/Dashboard/Dashboard.jsx
@@ -11,11 +11,11 @@ import { FaRegEdit } from 'react-icons/fa';
 export default function Dashboard() {
     const [pending, setPending] = useState(true);
     useEffect(() => {
-		const timeout = setTimeout(() => {
-			setPending(false);
-		}, 2000);
-		return () => clearTimeout(timeout);
-	}, []);
+        const timeout = setTimeout(() => {
+            setPending(false);
+        }, 2000);
+        return () => clearTimeout(timeout);
+    }, []);
 
     const dispatch = useDispatch()
 
@@ -45,9 +45,9 @@ export default function Dashboard() {
         },
         {
             name: "Actions",
-            cell: row => (<div className="actions"> <button type="button"><FaRegEdit /></button><button type="button" onClick={() => handleDeleteProduct(row)}><MdDeleteForever /></button></div>)
+            cell: row => (<div className="actions"> <NavLink to={`admin/edit/${row._id}`} type="button"><FaRegEdit /></NavLink><button type="button" onClick={() => handleDeleteProduct(row)}><MdDeleteForever /></button></div>)
         },
-        
+
         /*  {
              name: "",
              selector: "image",
@@ -55,13 +55,14 @@ export default function Dashboard() {
          } */
     ]
 
-    useEffect(()=>{
+    useEffect(() => {
         setItems(products)
     }, [products])
 
     const handleDeleteProduct = (row) => {
-        dispatch(deleteProduct(row._id));
-      }
+        if (window.confirm("Are you sure you want to remove this product??"))
+            dispatch(deleteProduct(row._id));
+    }
 
     return (
         <>
@@ -77,14 +78,13 @@ export default function Dashboard() {
                         title="My products"
                         striped
                         highlightOnHover
-                        pointerOnHover
-                        paginationPerPage = {5}
-                        paginationRowsPerPageOptions = {[5,8]}
+                        paginationPerPage={5}
+                        paginationRowsPerPageOptions={[5, 8]}
                         pagination
                     />
                 </div>
                 {/* <div className="create"> */}
-                    <NavLink className="create" to="/create" className="add-button">Create</NavLink>
+                <NavLink className="create" to="/create" className="add-button">Create</NavLink>
                 {/* </div> */}
             </div>
         </>
